+++ CHANGELOG.rst
@@ -4,6 +4,12 @@ ChangeLog
 
 0.1
 ===
+0.1.9
+-----
+- Fix default_connection when upgrade
+- Find default app instead of default.
+- Diff MySQL ddl.
+- Check tortoise config.
 
 0.1.8
 -----